Currently hacky but graph shows up
Added the graph view which currently will only load when the view is
clicked (this probably has something to do with the order in which
things are loaded). Therefore the graph will only be generated when the
simulation button is pressed. Further investigation into what is going
on here and how to make it run smoothly will have to be done.