Graphing works! Proof of concept complete
The graphing view works and is able to graph the data passed to it.
Right now the simulation looks hideous so the coming commits will be
primarily to make it look a lot nice. Will be starting by adding
bootstrap and then will work to try to implement a reactive surface
like they did in the meteor seminar